backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
requests==2.31.0
# Analytics and caching dependencies
redis==4.6.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
# Background task processing with Redis
celery[redis]==5.3.4
# Time series analysis
statsmodels==0.14.0
# Scientific computing for forecasting
scipy==1.11.4
# Performance monitoring
psutil==5.9.6
# Report scheduling
schedule==1.2.0
# Report export formats
reportlab==4.0.7
openpyxl==3.1.2
# Email functionality
email-validator==2.1.0
# Additional utilities
python-dateutil==2.8.2
# Excel chart support (optional)
xlsxwriter==3.1.9
# Backup encryption
cryptography>=41.0.0
# Off-site storage support
boto3==1.34.0
botocore==1.34.0
# YAML parsing for Docker Compose tests
PyYAML==6.0.1
# Image processing dependencies
Pillow==10.1.0
aiofiles==23.2.1
# OAuth2 and HTTP client dependencies
httpx==0.25.2
pydantic-settings==2.1.0